ProfileHandler can now AddNewProfile.
AddNewProfile uses GenerateSalt and HashPassword for security.
I have updated the database and ProfileModel.cs to contain the salt

BUG:
The database is copied to each project in the solution when i run it.
This makes the database out of sync.
For now, i have added the database to my server, so i can code.
I will fix the problem later on.