Changed the ProfileHandler to a singleton.
This is to prevent multiple people beeing logged in at the same time.
Added enums for profilestatus and profilelevel - to make the code easier to read
Added a check when logging in, to prevent multiple logging in when someone else allready are.
Added a message to the exception that DeleteProfile can throw.
Added a LogOut method
Added a FailedLogOut exception