Updated database.
Added two columms to Profiles - ProfilesNewPassword and ProfilesNewEmail
Updated SendEmail to contain the correct url to the new api

Minor fixes of UpdateProfile
Created web api (in a seperate solution. Since it contains the password to my CP)
Adding new password/email confirmation before change in database.
Added debug code to UpdateProfile
Changed the GenerateSalt method,
It no longer converts the byte array to a string, but instead returns it as a string.
This is to prevent build in sql injection protection from removing vital parts of the salt as it is passed around as a sql command parameter.