Initialized MessageHandler branch
MessageHandler.cs:
Initialized an internal HttpClient that all the methods use.
Added the MediaType DefaultRequestHeader to Json.
CreateFeedback():
Instanciates a FeedbackModel, populates it with the parameters given (rating, feedbackText).
Throws an exception if rating parameter is not between 0-11.
Sends a POST command to the WebAPI using Json the WebAPI doesn't know which table to post it to.
Removed NotImplementedException