DOES NOT COMPILE
This is my first shot at a e-mail sending solution. Added a e-mail NuGet package to both Phone and Windows 8.1 projects but it seems that it doesn't work with Store App. The Store version of this NuGet package needs it to be awaitable.

Question is if we should go with a nuget package or enable silverlight and use Morten's prototype solution (can be found in ProfileHandler.cs in the ProfileHandlerPrototype fork)