SendEmail():
I am putting this implementation on hold. It really should be server-side and implementing it is quite complicated and might need some external services. If you think it is important please contact Lárus Þór

Removed the external files/nuget packages I was trying for e-mail services.

Added a comment to all the catches to easily find all the cases where the following TODO needs to be done:

TODO: Put in error handling (logging, user messaging and other) in the catch statements.