MessageHandler.cs
GetNotifications():
Changed it so that it only returns Notifications who are to the profile currently logged in.

LandingPageViewModel.cs
Created
refreshMarketplace():
Name speaks for itself but throws a NotImplemented Exception because the bastard Bruno hasn't merged his TaskHandler so I can't access TaskHandler.GetTasks()
refreshActiveTasks():
This is just for the Active Tasks List on the landing page so it looks up all tasks in the local Marketplace collection and adds them to another collection that will be shown in the listview
refreshNotifications():
Notifications are going to be accessable in every page except the login page so there shouldn't be any logic aside from MessageHandler.GetNotifications(). Haven't been able to test if it works yet. I'm checking if the currentlogged in profile is the same as the one in the Notification ToProfile. But maybe it's better to check currentLoggedInProfile.Id with Notification FK_UserTo.

LandingPage.xaml:
Made the grid so you have 4 equally sized boxes in the middle and margin grids on the outside. Then inside the grids we have some stackpanels, some vertical, others horizontal.